backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for SmartTutor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.7",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.40.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "requests>=2.31.0",
    "langchain-groq>=1.1.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
